+++ src/providers.ts
@@ -47,14 +47,12 @@ import {
   OllamaCompletionProvider,
   OllamaChatProvider,
 } from './providers/ollama';
-import {
-  OpenAiCompletionProvider,
-  OpenAiChatCompletionProvider,
-  OpenAiEmbeddingProvider,
-  OpenAiModerationProvider,
-} from './providers/openai';
 import { OpenAiAssistantProvider } from './providers/openai/assistant';
+import { OpenAiChatCompletionProvider } from './providers/openai/chat';
+import { OpenAiCompletionProvider } from './providers/openai/completion';
+import { OpenAiEmbeddingProvider } from './providers/openai/embedding';
 import { OpenAiImageProvider } from './providers/openai/image';
+import { OpenAiModerationProvider } from './providers/openai/moderation';
 import { parsePackageProvider } from './providers/packageParser';
 import { PortkeyChatCompletionProvider } from './providers/portkey';
 import { PythonProvider } from './providers/pythonCompletion';
