+++ src/providers/azureUtil.ts
@@ -3,8 +3,9 @@ import { MODEL_GRADED_ASSERTION_TYPES } from '../assertions';
 import logger from '../logger';
 import type { TestCase, TestSuite } from '../types';
 import { AzureChatCompletionProvider, AzureCompletionProvider } from './azure';
-import { OpenAiChatCompletionProvider, OpenAiCompletionProvider } from './openai';
 import { OpenAiAssistantProvider } from './openai/assistant';
+import { OpenAiChatCompletionProvider } from './openai/chat';
+import { OpenAiCompletionProvider } from './openai/completion';
 
 export function maybeEmitAzureOpenAiWarning(testSuite: TestSuite, tests: TestCase[]) {
   const hasAzure = testSuite.providers.some(
