+++ src/providers/openai/index.ts
@@ -1,139 +1,14 @@
 import type OpenAI from 'openai';
-import { fetchWithCache } from '../../cache';
-import { getEnvString, getEnvFloat, getEnvInt } from '../../envars';
-import logger from '../../logger';
+import { getEnvString } from '../../envars';
 import type {
   ApiProvider,
   CallApiContextParams,
   CallApiOptionsParams,
-  ProviderEmbeddingResponse,
   ProviderResponse,
 } from '../../types';
 import type { EnvOverrides } from '../../types/env';
-import { renderVarsInObject } from '../../util';
-import { maybeLoadFromExternalFile } from '../../util';
-import { safeJsonStringify } from '../../util/json';
-import { calculateCost, REQUEST_TIMEOUT_MS, parseChatPrompt } from '../shared';
 import type { OpenAiSharedOptions } from './types';
-import { getTokenUsage, type OpenAiFunction, type OpenAiTool } from './util';
-
-// see https://platform.openai.com/docs/models
-export const OPENAI_CHAT_MODELS = [
-  ...['chatgpt-4o-latest'].map((model) => ({
-    id: model,
-    cost: {
-      input: 5 / 1e6,
-      output: 15 / 1e6,
-    },
-  })),
-  ...['o1', 'o1-2024-12-17', 'o1-preview', 'o1-preview-2024-09-12'].map((model) => ({
-    id: model,
-    cost: {
-      input: 15 / 1e6,
-      output: 60 / 1e6,
-    },
-  })),
-  ...['o1-mini', 'o1-mini-2024-09-12'].map((model) => ({
-    id: model,
-    cost: {
-      input: 3 / 1e6,
-      output: 12 / 1e6,
-    },
-  })),
-  ...['o3-mini', 'o3-mini-2025-01-31'].map((model) => ({
-    id: model,
-    cost: {
-      input: 1.1 / 1e6,
-      output: 4.4 / 1e6,
-    },
-  })),
-  ...['gpt-4o', 'gpt-4o-2024-11-20', 'gpt-4o-2024-08-06'].map((model) => ({
-    id: model,
-    cost: {
-      input: 2.5 / 1e6,
-      output: 10 / 1e6,
-    },
-  })),
-  ...['gpt-4o-2024-05-13'].map((model) => ({
-    id: model,
-    cost: {
-      input: 5 / 1000000,
-      output: 15 / 1000000,
-    },
-  })),
-  ...['gpt-4o-mini', 'gpt-4o-mini-2024-07-18'].map((model) => ({
-    id: model,
-    cost: {
-      input: 0.15 / 1000000,
-      output: 0.6 / 1000000,
-    },
-  })),
-  ...['gpt-4', 'gpt-4-0613'].map((model) => ({
-    id: model,
-    cost: {
-      input: 30 / 1000000,
-      output: 60 / 1000000,
-    },
-  })),
-  ...[
-    'gpt-4-turbo',
-    'gpt-4-turbo-2024-04-09',
-    'gpt-4-turbo-preview',
-    'gpt-4-0125-preview',
-    'gpt-4-1106-preview',
-  ].map((model) => ({
-    id: model,
-    cost: {
-      input: 10 / 1000000,
-      output: 30 / 1000000,
-    },
-  })),
-  {
-    id: 'gpt-3.5-turbo',
-    cost: {
-      input: 0.5 / 1000000,
-      output: 1.5 / 1000000,
-    },
-  },
-  {
-    id: 'gpt-3.5-turbo-0125',
-    cost: {
-      input: 0.5 / 1000000,
-      output: 1.5 / 1000000,
-    },
-  },
-  {
-    id: 'gpt-3.5-turbo-1106',
-    cost: {
-      input: 1 / 1000000,
-      output: 2 / 1000000,
-    },
-  },
-  ...['gpt-3.5-turbo-instruct'].map((model) => ({
-    id: model,
-    cost: {
-      input: 1.5 / 1000000,
-      output: 2 / 1000000,
-    },
-  })),
-];
-
-// See https://platform.openai.com/docs/models/model-endpoint-compatibility
-export const OPENAI_COMPLETION_MODELS = [
-  {
-    id: 'gpt-3.5-turbo-instruct',
-    cost: {
-      input: 1.5 / 1000000,
-      output: 2 / 1000000,
-    },
-  },
-  {
-    id: 'text-davinci-002',
-  },
-  {
-    id: 'text-babbage-002',
-  },
-];
+import { type OpenAiFunction, type OpenAiTool } from './util';
 
 export type OpenAiCompletionOptions = OpenAiSharedOptions & {
   temperature?: number;
@@ -254,440 +129,3 @@ export class OpenAiGenericProvider implements ApiProvider {
     throw new Error('Not implemented');
   }
 }
-
-export class OpenAiEmbeddingProvider extends OpenAiGenericProvider {
-  async callEmbeddingApi(text: string): Promise<ProviderEmbeddingResponse> {
-    if (!this.getApiKey()) {
-      throw new Error('OpenAI API key must be set for similarity comparison');
-    }
-
-    const body = {
-      input: text,
-      model: this.modelName,
-    };
-    let data,
-      cached = false;
-    try {
-      ({ data, cached } = (await fetchWithCache(
-        `${this.getApiUrl()}/embeddings`,
-        {
-          method: 'POST',
-          headers: {
-            'Content-Type': 'application/json',
-            Authorization: `Bearer ${this.getApiKey()}`,
-            ...(this.getOrganization() ? { 'OpenAI-Organization': this.getOrganization() } : {}),
-            ...this.config.headers,
-          },
-          body: JSON.stringify(body),
-        },
-        REQUEST_TIMEOUT_MS,
-      )) as unknown as any);
-    } catch (err) {
-      logger.error(`API call error: ${err}`);
-      throw err;
-    }
-    logger.debug(`\tOpenAI embeddings API response: ${JSON.stringify(data)}`);
-
-    try {
-      const embedding = data?.data?.[0]?.embedding;
-      if (!embedding) {
-        throw new Error('No embedding found in OpenAI embeddings API response');
-      }
-      return {
-        embedding,
-        tokenUsage: getTokenUsage(data, cached),
-      };
-    } catch (err) {
-      logger.error(data.error.message);
-      throw err;
-    }
-  }
-}
-
-export function formatOpenAiError(data: {
-  error: { message: string; type?: string; code?: string };
-}): string {
-  let errorMessage = `API error: ${data.error.message}`;
-  if (data.error.type) {
-    errorMessage += `, Type: ${data.error.type}`;
-  }
-  if (data.error.code) {
-    errorMessage += `, Code: ${data.error.code}`;
-  }
-  errorMessage += '\n\n' + safeJsonStringify(data, true /* prettyPrint */);
-  return errorMessage;
-}
-
-export function calculateOpenAICost(
-  modelName: string,
-  config: OpenAiSharedOptions,
-  promptTokens?: number,
-  completionTokens?: number,
-): number | undefined {
-  return calculateCost(modelName, config, promptTokens, completionTokens, [
-    ...OPENAI_CHAT_MODELS,
-    ...OPENAI_COMPLETION_MODELS,
-  ]);
-}
-
-export class OpenAiCompletionProvider extends OpenAiGenericProvider {
-  static OPENAI_COMPLETION_MODELS = OPENAI_COMPLETION_MODELS;
-
-  static OPENAI_COMPLETION_MODEL_NAMES = OPENAI_COMPLETION_MODELS.map((model) => model.id);
-
-  config: OpenAiCompletionOptions;
-
-  constructor(
-    modelName: string,
-    options: { config?: OpenAiCompletionOptions; id?: string; env?: EnvOverrides } = {},
-  ) {
-    super(modelName, options);
-    this.config = options.config || {};
-    if (
-      !OpenAiCompletionProvider.OPENAI_COMPLETION_MODEL_NAMES.includes(modelName) &&
-      this.getApiUrl() === this.getApiUrlDefault()
-    ) {
-      logger.warn(`FYI: Using unknown OpenAI completion model: ${modelName}`);
-    }
-  }
-
-  async callApi(
-    prompt: string,
-    context?: CallApiContextParams,
-    callApiOptions?: CallApiOptionsParams,
-  ): Promise<ProviderResponse> {
-    if (!this.getApiKey()) {
-      throw new Error(
-        'OpenAI API key is not set. Set the OPENAI_API_KEY environment variable or add `apiKey` to the provider config.',
-      );
-    }
-
-    let stop: string;
-    try {
-      stop = getEnvString('OPENAI_STOP')
-        ? JSON.parse(getEnvString('OPENAI_STOP') || '')
-        : this.config?.stop || ['<|im_end|>', '<|endoftext|>'];
-    } catch (err) {
-      throw new Error(`OPENAI_STOP is not a valid JSON string: ${err}`);
-    }
-    const body = {
-      model: this.modelName,
-      prompt,
-      seed: this.config.seed,
-      max_tokens: this.config.max_tokens ?? getEnvInt('OPENAI_MAX_TOKENS', 1024),
-      temperature: this.config.temperature ?? getEnvFloat('OPENAI_TEMPERATURE', 0),
-      top_p: this.config.top_p ?? getEnvFloat('OPENAI_TOP_P', 1),
-      presence_penalty: this.config.presence_penalty ?? getEnvFloat('OPENAI_PRESENCE_PENALTY', 0),
-      frequency_penalty:
-        this.config.frequency_penalty ?? getEnvFloat('OPENAI_FREQUENCY_PENALTY', 0),
-      best_of: this.config.best_of ?? getEnvInt('OPENAI_BEST_OF', 1),
-      ...(callApiOptions?.includeLogProbs ? { logprobs: callApiOptions.includeLogProbs } : {}),
-      ...(stop ? { stop } : {}),
-      ...(this.config.passthrough || {}),
-    };
-
-    logger.debug(`Calling OpenAI API: ${JSON.stringify(body)}`);
-    let data,
-      cached = false;
-    try {
-      ({ data, cached } = (await fetchWithCache(
-        `${this.getApiUrl()}/completions`,
-        {
-          method: 'POST',
-          headers: {
-            'Content-Type': 'application/json',
-            Authorization: `Bearer ${this.getApiKey()}`,
-            ...(this.getOrganization() ? { 'OpenAI-Organization': this.getOrganization() } : {}),
-            ...this.config.headers,
-          },
-          body: JSON.stringify(body),
-        },
-        REQUEST_TIMEOUT_MS,
-      )) as unknown as any);
-    } catch (err) {
-      logger.error(`API call error: ${String(err)}`);
-      return {
-        error: `API call error: ${String(err)}`,
-      };
-    }
-    logger.debug(`\tOpenAI completions API response: ${JSON.stringify(data)}`);
-    if (data.error) {
-      return {
-        error: formatOpenAiError(data),
-      };
-    }
-    try {
-      return {
-        output: data.choices[0].text,
-        tokenUsage: getTokenUsage(data, cached),
-        cached,
-        cost: calculateOpenAICost(
-          this.modelName,
-          this.config,
-          data.usage?.prompt_tokens,
-          data.usage?.completion_tokens,
-        ),
-      };
-    } catch (err) {
-      return {
-        error: `API error: ${String(err)}: ${JSON.stringify(data)}`,
-      };
-    }
-  }
-}
-
-export class OpenAiChatCompletionProvider extends OpenAiGenericProvider {
-  static OPENAI_CHAT_MODELS = OPENAI_CHAT_MODELS;
-
-  static OPENAI_CHAT_MODEL_NAMES = OPENAI_CHAT_MODELS.map((model) => model.id);
-
-  config: OpenAiCompletionOptions;
-
-  constructor(
-    modelName: string,
-    options: { config?: OpenAiCompletionOptions; id?: string; env?: EnvOverrides } = {},
-  ) {
-    if (!OpenAiChatCompletionProvider.OPENAI_CHAT_MODEL_NAMES.includes(modelName)) {
-      logger.debug(`Using unknown OpenAI chat model: ${modelName}`);
-    }
-    super(modelName, options);
-    this.config = options.config || {};
-  }
-
-  protected isReasoningModel(): boolean {
-    return this.modelName.startsWith('o1') || this.modelName.startsWith('o3');
-  }
-
-  protected supportsTemperature(): boolean {
-    // OpenAI's o1 and o3 models don't support temperature but some 3rd
-    // party reasoning models do.
-    return !this.isReasoningModel();
-  }
-
-  getOpenAiBody(
-    prompt: string,
-    context?: CallApiContextParams,
-    callApiOptions?: CallApiOptionsParams,
-  ) {
-    // Merge configs from the provider and the prompt
-    const config = {
-      ...this.config,
-      ...context?.prompt?.config,
-    };
-
-    const messages = parseChatPrompt(prompt, [{ role: 'user', content: prompt }]);
-
-    const isReasoningModel = this.isReasoningModel();
-    const maxCompletionTokens = isReasoningModel
-      ? (config.max_completion_tokens ?? getEnvInt('OPENAI_MAX_COMPLETION_TOKENS'))
-      : undefined;
-    const maxTokens = isReasoningModel
-      ? undefined
-      : (config.max_tokens ?? getEnvInt('OPENAI_MAX_TOKENS', 1024));
-
-    const temperature = this.supportsTemperature()
-      ? (config.temperature ?? getEnvFloat('OPENAI_TEMPERATURE', 0))
-      : undefined;
-    const reasoningEffort = isReasoningModel
-      ? renderVarsInObject(config.reasoning_effort, context?.vars)
-      : undefined;
-
-    const body = {
-      model: this.modelName,
-      messages,
-      seed: config.seed,
-      ...(maxTokens ? { max_tokens: maxTokens } : {}),
-      ...(maxCompletionTokens ? { max_completion_tokens: maxCompletionTokens } : {}),
-      ...(reasoningEffort ? { reasoning_effort: reasoningEffort } : {}),
-      ...(temperature ? { temperature } : {}),
-      ...(config.top_p !== undefined || process.env.OPENAI_TOP_P
-        ? { top_p: config.top_p ?? Number.parseFloat(process.env.OPENAI_TOP_P || '1') }
-        : {}),
-      ...(config.presence_penalty !== undefined || process.env.OPENAI_PRESENCE_PENALTY
-        ? {
-            presence_penalty:
-              config.presence_penalty ??
-              Number.parseFloat(process.env.OPENAI_PRESENCE_PENALTY || '0'),
-          }
-        : {}),
-      ...(config.frequency_penalty !== undefined || process.env.OPENAI_FREQUENCY_PENALTY
-        ? {
-            frequency_penalty:
-              config.frequency_penalty ??
-              Number.parseFloat(process.env.OPENAI_FREQUENCY_PENALTY || '0'),
-          }
-        : {}),
-      ...(config.functions
-        ? {
-            functions: maybeLoadFromExternalFile(
-              renderVarsInObject(config.functions, context?.vars),
-            ),
-          }
-        : {}),
-      ...(config.function_call ? { function_call: config.function_call } : {}),
-      ...(config.tools
-        ? { tools: maybeLoadFromExternalFile(renderVarsInObject(config.tools, context?.vars)) }
-        : {}),
-      ...(config.tool_choice ? { tool_choice: config.tool_choice } : {}),
-      ...(config.tool_resources ? { tool_resources: config.tool_resources } : {}),
-      ...(config.response_format
-        ? {
-            response_format: maybeLoadFromExternalFile(
-              renderVarsInObject(config.response_format, context?.vars),
-            ),
-          }
-        : {}),
-      ...(callApiOptions?.includeLogProbs ? { logprobs: callApiOptions.includeLogProbs } : {}),
-      ...(config.stop ? { stop: config.stop } : {}),
-      ...(config.passthrough || {}),
-    };
-
-    return { body, config };
-  }
-
-  async callApi(
-    prompt: string,
-    context?: CallApiContextParams,
-    callApiOptions?: CallApiOptionsParams,
-  ): Promise<ProviderResponse> {
-    if (!this.getApiKey()) {
-      throw new Error(
-        'OpenAI API key is not set. Set the OPENAI_API_KEY environment variable or add `apiKey` to the provider config.',
-      );
-    }
-
-    const { body, config } = this.getOpenAiBody(prompt, context, callApiOptions);
-    logger.debug(`Calling OpenAI API: ${JSON.stringify(body)}`);
-
-    let data, status, statusText;
-    let cached = false;
-    try {
-      ({ data, cached, status, statusText } = await fetchWithCache(
-        `${this.getApiUrl()}/chat/completions`,
-        {
-          method: 'POST',
-          headers: {
-            'Content-Type': 'application/json',
-            Authorization: `Bearer ${this.getApiKey()}`,
-            ...(this.getOrganization() ? { 'OpenAI-Organization': this.getOrganization() } : {}),
-            ...config.headers,
-          },
-          body: JSON.stringify(body),
-        },
-        REQUEST_TIMEOUT_MS,
-      ));
-
-      if (status < 200 || status >= 300) {
-        return {
-          error: `API error: ${status} ${statusText}\n${typeof data === 'string' ? data : JSON.stringify(data)}`,
-        };
-      }
-    } catch (err) {
-      logger.error(`API call error: ${String(err)}`);
-      return {
-        error: `API call error: ${String(err)}`,
-      };
-    }
-
-    logger.debug(`\tOpenAI chat completions API response: ${JSON.stringify(data)}`);
-    if (data.error) {
-      return {
-        error: formatOpenAiError(data),
-      };
-    }
-    try {
-      const message = data.choices[0].message;
-      if (message.refusal) {
-        return {
-          output: message.refusal,
-          tokenUsage: getTokenUsage(data, cached),
-          isRefusal: true,
-        };
-      }
-      let output = '';
-      if (message.reasoning) {
-        output = message.reasoning;
-      } else if (message.content && (message.function_call || message.tool_calls)) {
-        if (Array.isArray(message.tool_calls) && message.tool_calls.length === 0) {
-          output = message.content;
-        }
-        output = message;
-      } else if (message.content === null) {
-        output = message.function_call || message.tool_calls;
-      } else {
-        output = message.content;
-      }
-      const logProbs = data.choices[0].logprobs?.content?.map(
-        (logProbObj: { token: string; logprob: number }) => logProbObj.logprob,
-      );
-
-      // Handle structured output
-      if (config.response_format?.type === 'json_schema' && typeof output === 'string') {
-        try {
-          output = JSON.parse(output);
-        } catch (error) {
-          logger.error(`Failed to parse JSON output: ${error}`);
-        }
-      }
-
-      // Handle function tool callbacks
-      const functionCalls = message.function_call ? [message.function_call] : message.tool_calls;
-      if (functionCalls && config.functionToolCallbacks) {
-        const results = [];
-        for (const functionCall of functionCalls) {
-          const functionName = functionCall.name || functionCall.function?.name;
-          if (config.functionToolCallbacks[functionName]) {
-            try {
-              const functionResult = await config.functionToolCallbacks[functionName](
-                functionCall.arguments || functionCall.function?.arguments,
-              );
-              results.push(functionResult);
-            } catch (error) {
-              logger.error(`Error executing function ${functionName}: ${error}`);
-            }
-          }
-        }
-        if (results.length > 0) {
-          return {
-            output: results.join('\n'),
-            tokenUsage: getTokenUsage(data, cached),
-            cached,
-            logProbs,
-            cost: calculateOpenAICost(
-              this.modelName,
-              config,
-              data.usage?.prompt_tokens,
-              data.usage?.completion_tokens,
-            ),
-          };
-        }
-      }
-
-      return {
-        output,
-        tokenUsage: getTokenUsage(data, cached),
-        cached,
-        logProbs,
-        cost: calculateOpenAICost(
-          this.modelName,
-          config,
-          data.usage?.prompt_tokens,
-          data.usage?.completion_tokens,
-        ),
-      };
-    } catch (err) {
-      return {
-        error: `API error: ${String(err)}: ${JSON.stringify(data)}`,
-      };
-    }
-  }
-}
-
-export const DefaultEmbeddingProvider = new OpenAiEmbeddingProvider('text-embedding-3-large');
-export const DefaultGradingProvider = new OpenAiChatCompletionProvider('gpt-4o-2024-05-13');
-export const DefaultGradingJsonProvider = new OpenAiChatCompletionProvider('gpt-4o-2024-05-13', {
-  config: {
-    response_format: { type: 'json_object' },
-  },
-});
-export const DefaultSuggestionsProvider = new OpenAiChatCompletionProvider('gpt-4o-2024-05-13');
