+++ src/providers/xai.ts
@@ -1,7 +1,7 @@
 import type { ApiProvider, ProviderOptions } from '../types';
 import type { EnvOverrides } from '../types/env';
 import invariant from '../util/invariant';
-import { OpenAiChatCompletionProvider } from './openai';
+import { OpenAiChatCompletionProvider } from './openai/chat';
 
 export function createXAIProvider(
   providerPath: string,
