+++ src/suggestions.ts
@@ -1,5 +1,5 @@
 import { SUGGEST_PROMPTS_SYSTEM_MESSAGE } from './prompts';
-import { DefaultSuggestionsProvider } from './providers/openai';
+import { DefaultSuggestionsProvider } from './providers/openai/defaults';
 import type { TokenUsage } from './types';
 
 interface GeneratePromptsOutput {
