+++ test/assertions/python.test.ts
@@ -1,6 +1,6 @@
 import * as path from 'path';
 import { runAssertion } from '../../src/assertions';
-import { OpenAiChatCompletionProvider } from '../../src/providers/openai';
+import { OpenAiChatCompletionProvider } from '../../src/providers/openai/chat';
 import { runPython } from '../../src/python/pythonUtils';
 import { runPythonCode } from '../../src/python/wrapper';
 import type { Assertion, AtomicTestCase, GradingResult } from '../../src/types';
