+++ test/providers.test.ts
@@ -3,7 +3,8 @@ import yaml from 'js-yaml';
 import path from 'path';
 import { loadApiProvider, loadApiProviders } from '../src/providers';
 import { HttpProvider } from '../src/providers/http';
-import { OpenAiChatCompletionProvider, OpenAiEmbeddingProvider } from '../src/providers/openai';
+import { OpenAiChatCompletionProvider } from '../src/providers/openai/chat';
+import { OpenAiEmbeddingProvider } from '../src/providers/openai/embedding';
 import { PythonProvider } from '../src/providers/pythonCompletion';
 import { ScriptCompletionProvider } from '../src/providers/scriptCompletion';
 import { WebSocketProvider } from '../src/providers/websocket';
@@ -13,7 +14,8 @@ jest.mock('fs');
 jest.mock('js-yaml');
 jest.mock('../src/fetch');
 jest.mock('../src/providers/http');
-jest.mock('../src/providers/openai');
+jest.mock('../src/providers/openai/chat');
+jest.mock('../src/providers/openai/embedding');
 jest.mock('../src/providers/pythonCompletion');
 jest.mock('../src/providers/scriptCompletion');
 jest.mock('../src/providers/websocket');
