+++ test/providers/alibaba.test.ts
@@ -3,7 +3,8 @@ import {
   AlibabaChatCompletionProvider,
   AlibabaEmbeddingProvider,
 } from '../../src/providers/alibaba';
-import { OpenAiChatCompletionProvider, OpenAiEmbeddingProvider } from '../../src/providers/openai';
+import { OpenAiChatCompletionProvider } from '../../src/providers/openai/chat';
+import { OpenAiEmbeddingProvider } from '../../src/providers/openai/embedding';
 import type { ProviderOptions } from '../../src/types';
 
 jest.mock('../../src/logger');
