+++ test/redteam/providers/iterativeTree.test.ts
@@ -1,6 +1,6 @@
 import { jest } from '@jest/globals';
 import { v4 as uuidv4 } from 'uuid';
-import type { OpenAiChatCompletionProvider } from '../../../src/providers/openai';
+import type { OpenAiChatCompletionProvider } from '../../../src/providers/openai/chat';
 import type { TreeSearchOutput } from '../../../src/redteam/providers/iterativeTree';
 import {
   checkIfOnTopic,
